Give the suggested title one spelling
The dialog's seed and the issue's fallback each assembled the same three
arguments by hand from three modules, so they agreed only by coincidence;
`suggestTitleFor` now spells the recipe once and both call sites read it,
which is the invariant clampTitle's docstring already claimed. The dialog
stops importing crash-detector to reproduce a value crash-report can hand
it, and `unwoundFramesOf` names the pairing the frame filter always takes.

Let the required-field warnings key off the field rather than a DOM id.
The getter packed each localize key into a row with a synthetic id that
the render then searched back out, so a typo in either copy of the id
would type-check, drop the aria-describedby link, and still draw the
warning. The ids are constants now, written once each.

Fold the two BT-frame regexes into one with the label as a capture: they
shared an address pattern that had to stay byte-identical or a class of
frame would silently stop matching.

Move the crashReason and unwoundFrames suites beside the module that
exports them, and update crash-detector's header, which still described
the file as marker detection only.
