+++ src/util/crash-detector.ts
@@ -1,7 +1,10 @@
 import { tagged } from "./log-line.js";
 
 /**
- * Crash detection over device log lines.
+ * How ESPHome's crash output is spelled: where a dump starts and ends,
+ * the shape of a decoded frame, and the annotations the handler puts on
+ * its own report — the per-frame `(backtrace)` / `(stack scan)` label and
+ * the `Reason:` line.
  *
  * The marker patterns are ported from the regexes ESPHome's own
  * stacktrace decoders key on (`esphome/components/{esp32,esp8266}/
@@ -11,8 +14,10 @@ import { tagged } from "./log-line.js";
  * ...). Matching runs against `normalizeLogLine` output so the same
  * pattern hits regardless of transport (raw UART bytes vs the
  * backend's `\033`-literal ANSI vs the dialog's timestamp prefix).
- * The logs dialog latches on the first hit, so on hot streams this is
- * one batch scan until a crash is seen and zero cost after.
+ * Detection is the hot half: the logs dialog latches on the first hit,
+ * so on hot streams that is one batch scan until a crash is seen and
+ * zero cost after. Everything below it runs once, on a bounded excerpt,
+ * when the user opens a report.
  */
 
 /**
@@ -110,12 +115,12 @@ export function latchCrashKind(
   return current === "live" || next === null ? current : next;
 }
 
-// The esp32 crash handler labels every stored frame: `(backtrace)` for one
-// the unwinder produced, `(stack scan)` for a word that merely looks like a
-// return address in stack memory. A scanned word decodes to whatever symbol
-// owns that address, so it can name a component the crash never entered.
-const STACK_SCAN_RE = /BT\d+:\s*(?:0x)?([0-9a-fA-F]{8})\b.*\(stack scan\)/;
-const BACKTRACE_FRAME_RE = /BT\d+:\s*(?:0x)?([0-9a-fA-F]{8})\b.*\(backtrace\)/;
+// The label the esp32 crash handler puts on every stored frame:
+// `(backtrace)` for one the unwinder produced, `(stack scan)` for a word
+// that merely looks like a return address in stack memory. A scanned word
+// decodes to whatever symbol owns that address, so it can name a component
+// the crash never entered.
+const BT_FRAME_RE = /BT\d+:\s*(?:0x)?([0-9a-fA-F]{8})\b.*\((backtrace|stack scan)\)/;
 
 // The handler's own verdict, `Reason: <type>` or `<type> - <detail>`
 // (`Task wdt`, `Fault - Store access fault`). Two crashes can share a frame
@@ -131,7 +136,8 @@ const WATCHDOG_TYPE_RE = /wdt/i;
 const TRAILING_CAUSE_RE = /\s*\([^)]*\)\s*$/;
 
 /** What the crash handler blamed, preferring its specific cause over the
- *  exception class; "" when the dump carries no reason. */
+ *  exception class; "" when the dump carries no reason, or names one the
+ *  handler itself couldn't decode. */
 export function crashReason(excerpt: string[]): string {
   for (const line of excerpt) {
     const match = CRASH_REASON_RE.exec(line);
@@ -155,10 +161,10 @@ export function unwoundFrames(excerpt: string[], decodedFrames: string[]): strin
   const scanned = new Set<string>();
   const unwound = new Set<string>();
   for (const line of excerpt) {
-    const scan = STACK_SCAN_RE.exec(line);
-    if (scan) scanned.add(scan[1].toLowerCase());
-    const frame = BACKTRACE_FRAME_RE.exec(line);
-    if (frame) unwound.add(frame[1].toLowerCase());
+    const match = BT_FRAME_RE.exec(line);
+    if (match) {
+      (match[2] === "backtrace" ? unwound : scanned).add(match[1].toLowerCase());
+    }
   }
   // A recursive frame's return address litters the stack, so the scan
   // re-finds one the unwinder already vouched for. Filtering on the address
